feat: adding the entrypoint to run the vite placeholder envs

+++ docker/entrypoint.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 set -e
 
-# Substitui todos os placeholders nos assets JS
 echo "Injecting runtime environment variables..."
+
 for file in /var/www/html/public/build/assets/*.js; do
+  [ -f "$file" ] || continue
   envsubst < "$file" > "$file.tmp"
   mv "$file.tmp" "$file"
 done
 
-echo "Starting Apache..."
+echo "Starting Supervisor..."
+
 exec "$@"
